requirements.txt
streamlit>=1.32.0
openai-whisper>=20231117
imageio-ffmpeg>=0.4.9
fpdf2>=2.7.9
requests>=2.31.0
torch>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.1
pyperclip>=1.8.2
